chore : goal, teamProgress 파일 분할
