+++ src/widgets/goal/ui/index.ts
@@ -1,2 +1,3 @@
 export * from './goal';
 export * from './no-goal';
+export * from './TeamProgress';
